requirements.txt
pytest==7.4.4
python-dotenv==1.0.1
selenium==4.17.2
webdriver_manager==4.0.1
